Add files via upload
please use this as a learning tool and follow best  academic practices and do not plagiarize. i am sharing my work to help others get a better understanding of the expectations throughout the course. Please Note this is only my section of the presentation. As other group members are talking through it, it would not be appropriate to publicize the full presentation without permission.